refactor: 코드리뷰 반영
- https://github.com/next-step/java-racingcar/pull/1626#pullrequestreview-545360462
