+++ src/test/java/step3/domain/CarsTest.java
@@ -3,13 +3,28 @@ package step3.domain;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import step3.service.RacingRule;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class CarsTest {
 
-  public static final int NUMBER_OF_CAR = 5;
+  private static final int NUMBER_OF_CAR = 5;
   private Cars cars;
+
+  private RacingRule createTestRacingRule(boolean canMove) {
+    return new RacingRule() {
+      @Override
+      public boolean canMove(int targetNumber) {
+        return canMove;
+      }
+
+      @Override
+      public int getTargetNumber() {
+        return 0;
+      }
+    };
+  }
 
   @BeforeEach
   void before() {
@@ -19,20 +34,22 @@ public class CarsTest {
   @DisplayName("모두 이동 가능한 조건이라, 이동한 경우")
   @Test
   public void canMoveIsTrue() {
-    cars.race(new TestAllTrueRacingRule());
+    cars.race(createTestRacingRule(true));
     assertThat(cars.getCars().size()).isEqualTo(NUMBER_OF_CAR);
     for (Car car : cars.getCars()) {
-      assertThat(car.getDistance()).isEqualTo(Car.INIT_POSITION + 1);
+      assertThat(car.getDistance()).isEqualTo(2);
     }
   }
 
   @DisplayName("모두 이동 가능하지 못한 조건이라, 이동 못한 경우")
   @Test
   public void canMoveIsFalse() {
-    cars.race(new TestAllFalseRacingRule());
+    cars.race(createTestRacingRule(false));
     assertThat(cars.getCars().size()).isEqualTo(NUMBER_OF_CAR);
     for (Car car : cars.getCars()) {
-      assertThat(car.getDistance()).isEqualTo(Car.INIT_POSITION);
+      assertThat(car.getDistance()).isEqualTo(1);
     }
   }
+
+
 }
