--- src/test/java/step3/domain/TestAllFalseRacingRule.java
@@ -1,16 +1,0 @@
-package step3.domain;
-
-import step3.RacingRule;
-
-public class TestAllFalseRacingRule implements RacingRule {
-
-  @Override
-  public boolean canMove(int targetNumber) {
-    return false;
-  }
-
-  @Override
-  public int getTargetNumber() {
-    return 0;
-  }
-}
