--- src/test/java/step3/domain/TestAllTrueRacingRule.java
@@ -1,16 +1,0 @@
-package step3.domain;
-
-import step3.RacingRule;
-
-public class TestAllTrueRacingRule implements RacingRule {
-
-  @Override
-  public boolean canMove(int targetNumber) {
-    return true;
-  }
-
-  @Override
-  public int getTargetNumber() {
-    return 0;
-  }
-}
